Re-implemented the settings module.
The settings are no longer read from an app-specific module, but rather
injected into the project's settings module.

This allows real unitttesting (because Django's `override_setttings` is working)
and seems cleaner.

The injection is done in the app's AppConfig.ready() method.

Because one of the settings is used as a default value for the models, an
utility function had to be added, to actually read this setting.